+++ src/commands/helps/publish.rs
@@ -23,6 +23,8 @@ Usage:
 
 Options:
     --ready, -y           run the operation to completion
+    --channel, -c <name>...
+                          use this channel during publishing
     --no-install          do not install the ip for future use
     --list, -l            view available channels and exit
     --all-public          publish with all source files being public