+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 - Adds `-p` switch for `--protocol` under `orbit install` command
 - Adds `-l` switch for `--list` flag in various subcommands
 - Allows default protocol to be set under `[install]` table in configuration file
+- Allows channels to be either set by default in configuration using `default-channels` in the `[publish]` section or set on the command-line using `--channel, -c` for the publish subcommand
 
 ### Changes
 - Renames `--no-clean` flag to `--keep`, `-k`, to align with conventional linux command-line options